+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/techknowlogic/BlueCarousalParkingAtWarehouse.java
@@ -8,18 +8,11 @@ import org.firstinspires.ftc.teamcode.drive.SampleMecanumDrive;
 
 @Autonomous
 @Config
-public class BlueCarousalParkingAtWarehouse extends BaseAutonomous {
-
-    public static double DRIVE_TO_HUB_STEP1_STRAFE_LEFT = 54;
-    public static double DRIVE_TO_HUB_STEP2_BACK = 3;
-
-    public static double DRIVE_TO_CAROUSAL_STEP1_STRAFE_RIGHT = 50;
-    public static double DRIVE_TO_CAROUSAL_STEP2_FORWARD = 40;
-    public static double DRIVE_TO_CAROUSAL_STEP3_STRAFE_RIGHT = 10;
+public class BlueCarousalParkingAtWarehouse extends BaseBlueCarousal {
 
     public static double PARK_ROBOT_STEP1_BACK = 35;
-    public static double PARK_ROBOT_STEP2_STRAFE_RIGHT = 10;
-    public static double PARK_ROBOT_STEP3_BACK = 90;
+    public static double PARK_ROBOT_STEP2_STRAFE_RIGHT = 20;
+    public static double PARK_ROBOT_STEP3_BACK = 95;
 
     @Override
     protected void parkRobot(SampleMecanumDrive driveTrain) {
@@ -37,56 +30,5 @@ public class BlueCarousalParkingAtWarehouse extends BaseAutonomous {
                 .back(PARK_ROBOT_STEP3_BACK)
                 .build();
         driveTrain.followTrajectory(step3Back);
-    }
-
-    @Override
-    protected int getElevatorLevel(String shippingElementPosition) {
-        if (shippingElementPosition.equals("LEFT")) {
-            return 1;
-        } else if (shippingElementPosition.equals("RIGHT")) {
-            return 2;
-        } else {
-            return 3;
-        }
-    }
-
-    @Override
-    protected void driveToShippingHub(SampleMecanumDrive driveTrain) {
-
-        driveTrain.turn(Math.toRadians(90));
-
-        Trajectory strafeLeft = driveTrain.trajectoryBuilder(driveTrain.getPoseEstimate(), false)
-                .strafeLeft(DRIVE_TO_HUB_STEP1_STRAFE_LEFT)
-                .build();
-        driveTrain.followTrajectory(strafeLeft);
-
-        Trajectory back = driveTrain.trajectoryBuilder(driveTrain.getPoseEstimate(), false)
-                .back(DRIVE_TO_HUB_STEP2_BACK)
-                .build();
-        driveTrain.followTrajectory(back);
-    }
-
-    @Override
-    protected boolean isCarousalSpinReversed() {
-        return true;
-    }
-
-    @Override
-    protected void driveToCarousal(SampleMecanumDrive driveTrain) {
-
-        Trajectory strafeRight = driveTrain.trajectoryBuilder(driveTrain.getPoseEstimate(), false)
-                .strafeRight(DRIVE_TO_CAROUSAL_STEP1_STRAFE_RIGHT)
-                .build();
-        driveTrain.followTrajectory(strafeRight);
-
-        Trajectory forward = driveTrain.trajectoryBuilder(driveTrain.getPoseEstimate(), false)
-                .forward(DRIVE_TO_CAROUSAL_STEP2_FORWARD)
-                .build();
-        driveTrain.followTrajectory(forward);
-
-        Trajectory step3StrafeRight = driveTrain.trajectoryBuilder(driveTrain.getPoseEstimate(), false)
-                .strafeRight(DRIVE_TO_CAROUSAL_STEP3_STRAFE_RIGHT)
-                .build();
-        driveTrain.followTrajectory(step3StrafeRight);
     }
 }
